+++ src/chris_package/CMakeLists.txt
@@ -1,28 +1,29 @@
 cmake_minimum_required(VERSION 3.8)
-project(yourname_package)
+project(chris_package)
 
 if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-Wall -Wextra -Wpedantic)
 endif()
 
 # find dependencies
+find_package(cv_bridge REQUIRED)
 find_package(ament_cmake REQUIRED)
-
 find_package(rclcpp REQUIRED)
 find_package(std_msgs REQUIRED)
-
+find_package(sensor_msgs REQUIRED)
 # uncomment the following section in order to fill in
 # further dependencies manually.
 # find_package(<dependency> REQUIRED)
 
-add_executable(yourname_node src/yourname_node.cpp)
-ament_target_dependencies(yourname_node rclcpp std_msgs)
-target_include_directories(yourname_node PUBLIC
+add_executable(chris_node src/chris.cpp)
+ament_target_dependencies(chris_node rclcpp sensor_msgs cv_bridge)
+target_include_directories(chris_node PUBLIC
   $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
   $<INSTALL_INTERFACE:include>)
-target_compile_features(yourname_node PUBLIC c_std_99 cxx_std_17)  # Require C99 and C++17
+target_compile_features(chris_node PUBLIC c_std_99 cxx_std_17)  # Require C99 and C++17
 
-install(TARGETS yourname_node
+
+install(TARGETS chris_node
   DESTINATION lib/${PROJECT_NAME})
 
 if(BUILD_TESTING)
